requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
Pillow==10.0.0
psycopg2-binary
